ODOL Truck Detection Web/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
huggingface-hub==0.21.4
ultralytics==8.1.24
tensorflow==2.20.0
torch==2.5.1
torchvision==0.20.1
